fix: restore mxDictionary map index signature parameter set to string
Fix error TS1023: An index signature parameter type must be either 'string' or
'number'. In addition, references to mxgraph javascript code is available to
explain why the key is a string.

`skipLibCheck` is now set to `false` in the tsconfig.json file. So, tsc run
prior checking the tests will fail in case of TS errors and we will be alerted
before integrating such problem in the repository.
Projects integrating typed-mxgraph will also be able to use this option without
errors coming from the mxgraph types.